♻️ Reorganize the sample
