+++ v2/src/core/data/types.ts
@@ -1,0 +1,8 @@
+export interface IEntity {
+  id: string;
+}
+
+export interface Storage<TEntity extends IEntity> {
+  getEntities(): Promise<ReadonlyArray<TEntity>>;
+  save(entity: TEntity): Promise<void>;
+}
