+++ v2/src/core/ioc/dependencyCollection.ts
