+++ v2/src/core/schema/types.ts
@@ -1,19 +1,3 @@
-export type HttpVerb = "get" | "post" | "patch";
-
-export interface IController<TModel = unknown> {
-  getAll(test: any): Promise<ReadonlyArray<Readonly<TModel>>>;
-  post(model: TModel): Promise<void>;
-}
-
-export interface IAction {
-  controller: string;
-  method: string;
-}
-
-export interface IRoute extends IAction {
-  httpVerb: HttpVerb;
-  path: string;
-}
 
 type SchemaFrom = 'params' | 'querystring' | 'body';
 
@@ -39,7 +23,7 @@ export interface Schema {
   [key: string]: SchemaPropertyOptions
 }
 
-export interface JsonSchemaProperty {
+interface JsonSchemaProperty {
   type: 'object',
   properties: { [key: string]: {
     type: 'integer' | 'string',
