--- v2/src/data/entity.ts
@@ -1,3 +1,0 @@
-export interface Entity {
-  id: string;
-}
