--- v2/src/data/storage.ts
@@ -1,6 +1,0 @@
-import { Entity } from "./entity";
-
-export interface Storage<TEntity extends Entity> {
-  getEntities(): Promise<ReadonlyArray<TEntity>>;
-  save(entity: TEntity): Promise<void>;
-}
